make visible UserTrening and UserDiet on the mane page

+++ FitnessApp_.Net/FitnessApp.DAL/Models/Meal.cs
@@ -21,7 +21,7 @@ namespace FitnessApp.DAL.Models
         public double Protein { get; set; }
         public double Fat { get; set; }
         public double Carbon { get; set; }
-        public int CalorificOfMeal { get; set; }
+        public double CalorificOfMeal { get; set; }
 
         public int TypeOfMealId { get; set; }
         public virtual TypeOfMeal TypeOfMeal { get; set; }
